# fern/apis/fai/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FAI",
    "summary": "The FAI API.",
    "version": "0.0.0"
  },
  "paths": {
...
    "/feedback/{domain}": {
      "post": {
        "tags": [
          "Feedback"
        ],
        "summary": "Create Feedback",
        "operationId": "create_feedback",
        "parameters": [
          {
            "name": "domain",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Domain"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFeedbackRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateFeedbackResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/feedback/{domain}/{conversation_id}": {
      "get": {
        "tags": [
          "Feedback"
        ],
        "summary": "Get Feedback By Id",
        "operationId": "get_feedback_by_id",
        "parameters": [
          {
            "name": "domain",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Domain"
            }
          },
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Conversation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetFeedbackResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateFeedbackRequest": {
        "properties": {
          "conversation_id": {
            "type": "string",
            "title": "Conversation Id",
            "description": "The ID of the conversation"
          },
          "query_id": {
            "type": "string",
            "title": "Query Id",
            "description": "The ID of the query"
          },
          "domain": {
            "type": "string",
            "title": "Domain",
            "description": "The domain of the conversation"
          },
          "is_helpful": {
            "type": "boolean",
            "title": "Is Helpful",
            "description": "Whether the conversation was helpful"
          },
          "feedback_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feedback Message",
            "description": "The feedback message from the user"
          },
          "user_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Email",
            "description": "The email of the user"
          }
        },
        "type": "object",
        "required": [
          "conversation_id",
          "query_id",
          "domain",
          "is_helpful",
          "feedback_message",
          "user_email"
        ],
        "title": "CreateFeedbackRequest"
      },
      "CreateFeedbackResponse": {
        "properties": {
          "feedback_id": {
            "type": "string",
            "title": "Feedback Id",
            "description": "The ID of the created feedback"
          }
        },
        "type": "object",
        "required": [
          "feedback_id"
        ],
        "title": "CreateFeedbackResponse"
      },
...
      "Feedback": {
        "properties": {
          "query_id": {
            "type": "string",
            "title": "Query Id"
          },
          "conversation_id": {
            "type": "string",
            "title": "Conversation Id"
          },
          "domain": {
            "type": "string",
            "title": "Domain"
          },
          "is_helpful": {
            "type": "boolean",
            "title": "Is Helpful"
          },
          "feedback_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feedback Message"
          },
          "user_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Email"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "query_id",
          "conversation_id",
          "domain",
          "is_helpful",
          "created_at"
        ],
        "title": "Feedback"
      },
...
      "GetFeedbackResponse": {
        "properties": {
          "feedback": {
            "$ref": "#/components/schemas/Feedback",
            "description": "The returned feedback"
          }
        },
        "type": "object",
        "required": [
          "feedback"
        ],
        "title": "GetFeedbackResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
  "servers": [
    {
      "url": "https://fai.buildwithfern.com",
      "x-fern-server-name": "Production"
    },
    {
      "url": "https://fai-dev.buildwithfern.com",
      "x-fern-server-name": "Development"
    },
    {
      "url": "http://localhost:8080",
      "x-fern-server-name": "Local"
    }
  ]
}