# packages/fai-sdk/src/api/types/Feedback.ts
/**
 * This file was auto-generated by Fern from our API Definition.
 */

export interface Feedback {
    query_id: string;
    conversation_id: string;
    domain: string;
    is_helpful: boolean;
    feedback_message?: string;
    user_email?: string;
    created_at: string;
}