# packages/fai-sdk/src/api/types/GetFeedbackResponse.ts
/**
 * This file was auto-generated by Fern from our API Definition.
 */

import * as FernAI from "../index.js";

export interface GetFeedbackResponse {
    /** The returned feedback */
    feedback: FernAI.Feedback;
}